fix(compiler): clean semantic utility ownership
