--- packages/compiler/packages/semantic-utils/src/instructions/moduleEnd.ts
@@ -1,14 +1,0 @@
-import type { CompilationContext, ModuleEndLine } from '@8f4e/language-spec';
-
-import { popBlock } from '../blockStack';
-
-/**
- * Closes the active module block after tokenizer placement has matched the block structure.
- *
- * @param _line - Unused source AST line kept for handler signature consistency.
- * @param context - Compilation context used by the operation.
- * @returns Nothing.
- */
-export default function semanticModuleEnd(_line: ModuleEndLine, context: CompilationContext) {
-	popBlock(context);
-}
