+++ packages/compiler/packages/semantic-utils/src/operandTypes.test.ts
@@ -1,11 +1,6 @@
 import type { StackItem } from '@8f4e/language-spec';
 import { describe, expect, it } from 'vitest';
-import {
-	areAllOperandsFloat64,
-	areAllOperandsFloats,
-	areAllOperandsIntegers,
-	hasMixedFloatWidth,
-} from './operandTypes';
+import { areAllOperandsFloat64, areAllOperandsIntegers } from './operandTypes';
 
 describe('operandTypes utilities', () => {
 	const intOperand: StackItem = { kind: 'value', valueType: 'int', isNonZero: false };
@@ -25,22 +20,6 @@ describe('operandTypes utilities', () => {
 
 		it('returns true for empty operands', () => {
 			expect(areAllOperandsIntegers()).toBe(true);
-		});
-	});
-
-	describe('areAllOperandsFloats', () => {
-		it('returns true when all operands are floats', () => {
-			expect(areAllOperandsFloats(floatOperand, floatOperand)).toBe(true);
-			expect(areAllOperandsFloats(floatOperand)).toBe(true);
-		});
-
-		it('returns false when at least one operand is integer', () => {
-			expect(areAllOperandsFloats(floatOperand, intOperand)).toBe(false);
-			expect(areAllOperandsFloats(intOperand)).toBe(false);
-		});
-
-		it('returns true for empty operands', () => {
-			expect(areAllOperandsFloats()).toBe(true);
 		});
 	});
 
@@ -65,30 +44,6 @@ describe('operandTypes utilities', () => {
 
 		it('returns false for empty operands', () => {
 			expect(areAllOperandsFloat64()).toBe(false);
-		});
-	});
-
-	describe('hasMixedFloatWidth', () => {
-		it('returns true when one operand is float32 and another is float64', () => {
-			expect(hasMixedFloatWidth(floatOperand, float64Operand)).toBe(true);
-			expect(hasMixedFloatWidth(float64Operand, floatOperand)).toBe(true);
-		});
-
-		it('returns false when all operands are float32', () => {
-			expect(hasMixedFloatWidth(floatOperand, floatOperand)).toBe(false);
-		});
-
-		it('returns false when all operands are float64', () => {
-			expect(hasMixedFloatWidth(float64Operand, float64Operand)).toBe(false);
-		});
-
-		it('returns false when all operands are integers', () => {
-			expect(hasMixedFloatWidth(intOperand, intOperand)).toBe(false);
-		});
-
-		it('returns false for single float operand', () => {
-			expect(hasMixedFloatWidth(floatOperand)).toBe(false);
-			expect(hasMixedFloatWidth(float64Operand)).toBe(false);
 		});
 	});
 });
