+++ packages/compiler/packages/semantic-utils/src/operandTypes.ts
@@ -11,16 +11,6 @@ export function areAllOperandsIntegers(...operands: StackItem[]): boolean {
 }
 
 /**
- * Returns whether every operand is any supported float stack value.
- *
- * @param operands - Stack operands to inspect.
- * @returns Whether the check succeeds.
- */
-export function areAllOperandsFloats(...operands: StackItem[]): boolean {
-	return operands.every(operand => operand.valueType === 'float' || operand.valueType === 'float64');
-}
-
-/**
  * Returns whether every operand is a float64 stack value and at least one operand exists.
  *
  * @param operands - Stack operands to inspect.
@@ -28,15 +18,4 @@ export function areAllOperandsFloats(...operands: StackItem[]): boolean {
  */
 export function areAllOperandsFloat64(...operands: StackItem[]): boolean {
 	return operands.length > 0 && operands.every(operand => operand.valueType === 'float64');
-}
-
-/**
- * Returns whether the operands mix float32 and float64 stack values.
- *
- * @param operands - Stack operands to inspect.
- * @returns Whether the check succeeds.
- */
-export function hasMixedFloatWidth(...operands: StackItem[]): boolean {
-	const floats = operands.filter(op => op.valueType === 'float' || op.valueType === 'float64');
-	return floats.some(op => op.valueType === 'float64') && floats.some(op => op.valueType === 'float');
 }
