+++ packages/compiler/packages/wasm-codegen/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@8f4e/compiler-wasm-utils": "*",
     "@8f4e/language-spec": "*",
+    "@8f4e/memory-reference-inliner": "*",
     "@8f4e/semantic-utils": "*",
     "@8f4e/stack-analyzer": "*"
   },
