+++ packages/compiler/packages/wasm-codegen/src/normalization/default.ts
