+++ packages/compiler/packages/wasm-codegen/src/normalization/index.ts
