+++ packages/compiler/packages/wasm-codegen/src/normalization/localVariable.ts
