+++ packages/compiler/packages/wasm-codegen/src/normalization/map.ts
