+++ packages/compiler/packages/wasm-codegen/src/normalization/memoryDeclaration.ts
