+++ packages/compiler/packages/wasm-codegen/src/normalizeValueArguments.test.ts
