+++ packages/compiler/packages/wasm-codegen/vite.config.ts
@@ -6,7 +6,13 @@ const baseConfig = createLibConfig({
 	outDir: 'dist',
 	formats: ['es'],
 	fileName: () => 'index.js',
-	external: ['@8f4e/compiler-wasm-utils', '@8f4e/language-spec', '@8f4e/semantic-utils', '@8f4e/stack-analyzer'],
+	external: [
+		'@8f4e/compiler-wasm-utils',
+		'@8f4e/language-spec',
+		'@8f4e/memory-reference-inliner',
+		'@8f4e/semantic-utils',
+		'@8f4e/stack-analyzer',
+	],
 });
 
 export default defineConfig({
